Move nexus.py to proper location, refactor
